requirements.txt
amqp==5.3.1
asgiref==3.9.1
async-timeout==5.0.1
billiard==4.2.1
celery==5.3.4
certifi==2025.7.14
charset-normalizer==3.4.2
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
Django==4.2.7
django-ratelimit==4.1.0
django-redis==5.4.0
djangorestframework==3.14.0
drf-yasg==1.21.7
env==0.1.0
exceptiongroup==1.3.0
factory-boy==3.3.0
Faker==37.4.2
freezegun==1.2.2
idna==3.10
inflection==0.5.1
iniconfig==2.1.0
kombu==5.5.4
packaging==25.0
pluggy==1.6.0
prompt_toolkit==3.0.51
psycopg2-binary==2.9.9
pytest==7.4.3
pytest-django
pytest-cov
pytest-mock
pytest-django==4.7.0
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2025.2
PyYAML==6.0.2
redis==5.0.1
requests==2.31.0
six==1.17.0
sqlparse==0.5.3
tomli==2.2.1
typing_extensions==4.14.1
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.13
